+++ InventorySystem/Controllers/MarcaController.cs
@@ -1,5 +1,6 @@
 ﻿using InventorySystem.Core.Application.Interface.Services;
 using InventorySystem.Core.Application.ViewModel.Marca;
+using InventorySystem.Middlewares;
 using Microsoft.AspNetCore.Mvc;
 
 namespace InventorySystem.Controllers
@@ -7,23 +8,47 @@ namespace InventorySystem.Controllers
     public class MarcaController : Controller
     {
         private readonly IMarcaService _marcaService;
-        public MarcaController(IMarcaService marcaService)
+        private readonly ValidateUserSession _validateUserSession;
+        public MarcaController(IMarcaService marcaService, ValidateUserSession validateUserSession)
         {
             _marcaService = marcaService;
+            _validateUserSession = validateUserSession;
         }
         public async Task<IActionResult> Index()
         {
+            if (!_validateUserSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+            }
             return View(await _marcaService.GetAllViewModel());
         }
 
         public IActionResult Create()
         {
+            if (!_validateUserSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+            }
+            if (!_validateUserSession.hasAdmin())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
+            }
             return View();
         }
 
         [HttpPost]
         public async Task<IActionResult> Create(MarcaSaveViewModel vm)
         {
+            if (!_validateUserSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+            }
+            if (!_validateUserSession.hasAdmin())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
+            }
             if (!ModelState.IsValid)
             {
                 return View(vm);
@@ -34,6 +59,15 @@ namespace InventorySystem.Controllers
 
         public async Task<IActionResult> Edit(int id)
         {
+            if (!_validateUserSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+            }
+            if (!_validateUserSession.hasAdmin())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
+            }
             var vm = await _marcaService.GetById(id);
             return View(vm);
         }
@@ -41,6 +75,15 @@ namespace InventorySystem.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(MarcaSaveViewModel vm)
         {
+            if (!_validateUserSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+            }
+            if (!_validateUserSession.hasAdmin())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
+            }
             if (!ModelState.IsValid)
             {
                 return View(vm);
@@ -51,12 +94,30 @@ namespace InventorySystem.Controllers
 
         public async Task<IActionResult> Delete(int id)
         {
+            if (!_validateUserSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+            }
+            if (!_validateUserSession.hasAdmin())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
+            }
             return View("Delete",await _marcaService.GetById(id));
         }
 
         [HttpPost]
         public async Task<IActionResult> DeletePost(int id)
         {
+            if (!_validateUserSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+            }
+            if (!_validateUserSession.hasAdmin())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
+            }
             await _marcaService.Delete(id);
             return RedirectToRoute(new { controller = "Marca", action = "Index" });
         }
